Move kumaAPI into the container under `api`
